+++ gems/aws-sdk-robomaker/lib/aws-sdk-robomaker.rb
@@ -42,6 +42,6 @@ require_relative 'aws-sdk-robomaker/customizations'
 # @service
 module Aws::RoboMaker
 
-  GEM_VERSION = '1.18.0'
+  GEM_VERSION = '1.19.0'
 
 end